+++ R/block.R
@@ -460,13 +460,30 @@ process_tangle.block = function(x) {
 }
 #' @export
 process_tangle.inline = function(x) {
-  if (opts_knit$get('documentation') == 2L) {
-    return(paste(line_prompt(x$input.src, "#' ", "#' "), collapse = '\n'))
+  .append <- function(x, y) {
+    if (!nzchar(y)) x else if (!nzchar(x)) y else
+      paste0(x, '\n', y)
   }
-  code = x$code
-  if (length(code) == 0L || !any(idx <- grepl('knit_child\\(.+\\)', code)))
-    return('')
-  paste(c(sapply(code[idx], function(z) eval(parse_only(z))), ''), collapse = '\n')
+
+  output <- ''
+
+  if (opts_knit$get('documentation') == 2L) {
+    output <- paste(line_prompt(x$input.src, "#' ", "#' "), collapse = '\n')
+  }
+
+  code <- x$code
+  if (length(code) == 0L) return(output)
+
+  if (getOption('knitr.purl.inline', FALSE)) output <- .append(output, code)
+
+  idx <- grepl('knit_child\\(.+\\)', code)
+  if (any(idx)) {
+    cout <- sapply(code[idx], function(z) eval(parse_only(z)))
+    cout <- paste(c(cout, ''), collapse = '\n')
+    output <- .append(output, cout)
+  }
+
+  output
 }