+++ NEWS.md
@@ -10,6 +10,8 @@
 
 - `purl()` no longer writes verbatim non-R code chunks to the output R script; these code chunks will be commented out in the output
 
+- `purl()` can now also retrieve the code from inline expressions. It is disabled by default for compatibility, and can be enabled via the `knitr.purl.inline` option (cf purl documentation).
+
 # CHANGES IN knitr VERSION 1.15.1
 
 ## NEW FEATURES